pyproject.toml
[project]
name = "ginkgo"
version = "0.1.0"
description = "A dynamic, reproducible workflow orchestrator for genomics"
requires-python = ">= 3.11"
dependencies = [
    "numpy>=2.0",
    "ortools>=9.11",
    "pandas>=2.0",
    "pyarrow>=23.0.1,<24",
    "pyyaml>=6.0",
    "rich>=13.9",
]

[project.scripts]
ginkgo = "ginkgo.cli:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/ginkgo"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
ginkgo = { path = ".", editable = true }
numpy = ">=2.0"
ortools = ">=9.11"
pandas = ">=2.0"
pyarrow = ">=18.0"
rich = ">=13.9"

[tool.pixi.dependencies]
python = ">=3.11"
pyyaml = ">=6.0"

[tool.pixi.feature.dev.dependencies]
pytest = ">=8.0"
ruff = ">=0.4"

[tool.pixi.environments]
default = { features = ["dev"], solve-group = "default" }

[tool.pixi.tasks]
test = "pytest tests/ -v"
lint = "ruff check src/ tests/"
format = "ruff format src/ tests/"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
src = ["src"]
line-length = 99
